Refactored API action specs as request specs
Necessary to simplify testing while also testing the full API stack. The only oddity is the API log create spec since we need to test logging so keeping this as a unit spec still makes sense.

Milestone: minor
